+++ solutions/ailyasov/2/sources/2_4/trade_msg.h
@@ -18,7 +18,8 @@ enum TYPES {
 
 class TradeMsg {
     public:
-        TradeMsg(const size_t type, const size_t time, const size_t len, const std::string msg) : 
+        TradeMsg(const unsigned int type, const unsigned int time, const
+                unsigned int len, const std::string msg) : 
             type_(type), time_(time), len_(len), msg_(msg) {
                 //empty
             } 
@@ -28,14 +29,14 @@ class TradeMsg {
         friend std::ostream& operator<<(std::ostream&, const TradeMsg&);
         static std::vector<TradeMsg> read_data(Reader&);
         static void write_data(std::vector<TradeMsg>, Writer&);
-        size_t get_type() const { return type_; }
-        size_t get_time() const { return time_; }
-        size_t get_len() const { return len_; }
+        unsigned int get_type() const { return type_; }
+        unsigned int get_time() const { return time_; }
+        unsigned int get_len() const { return len_; }
         std::string get_msg() const { return msg_; }
-        static size_t max_time;
+        static unsigned int max_time;
     private:
-        size_t type_;
-        size_t time_;
-        size_t len_;
+        unsigned int type_;
+        unsigned int time_;
+        unsigned int len_;
         std::string msg_;
 };
